[ADD] Added sale_product_smart_buttons module
