--- invoice_product_smart_buttons/views/invoice_line_view.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="utf-8" ?>
-<odoo>
-
-	<!--  Create Action for smart buttons -->
-	<record id="action_product_invoice_list" model="ir.actions.act_window">
-		<field name="name">Invoice Lines</field>
-		<field name="res_model">account.invoice.line</field>
-		<field name="context">{'search_default_product_id': [active_id], 'default_product_id': active_id}</field>
-		<field name="domain">[('invoice_id.state', 'in', ['open', 'done'])]</field>
-	</record>
-	
-</odoo>
