+++ sale_product_smart_buttons/LICENSE/LICENSE
