+++ sale_product_smart_buttons/static/description/index.html
@@ -2,12 +2,12 @@
     <div class="oe_row oe_spaced" style="background:#dcdcdc;padding-top:2px;padding-bottom:25px;">
         <div class="oe_span12">
         <p class='oe_mt32 text-center' style="margin-left:60px;margin-right:60px;font-size:18px;font-family: Times New Roman">
-            This module adds 2 Smart buttons to show invoiced Qty and amount for particular product.
+            This module adds a Smart button to show total Sales amount for particular product.
             </p>
         </div> 
         <div class="oe_span12">
             <div class="oe_demo oe_picture oe_screenshot">
-                    <img src="invoice_amnt_qty_button.png">
+                    <img src="sales_samrt_button.png">
             </div>
         </div>
     </div>
